ArchivedReports logic now working. Refactored files. Multiple css changes.
