+++ src/AddAndEdit/AddStory.js
@@ -36,7 +36,7 @@ const AddStory = () => {
     };
 
     return (
-        <main className="p-10 bg-zinc-200">
+        <main className="h-screen p-10 bg-zinc-200">
             <h1 className="grid text-gray-800 text-4xl place-content-center font-semibold underline underline-offset-8 decoration-1 mb-10 mt-16">Add Story</h1>
             <div className="mt-10 sm:mt-0">
                 <div className="md:grid md:grid-cols-3 md:gap-6">
