+++ src/Components/GalleryPage.js
@@ -1,9 +1,9 @@
 import { useEffect, useState } from "react";
 import { listAll, ref, getDownloadURL } from "firebase/storage";
 import { collection, query, where, onSnapshot } from "firebase/firestore";
-import { storage, auth, db } from "./firebase";
+import { storage, auth, db } from "../firebase";
 import { useNavigate } from "react-router-dom";
-import SkeletonImage from "./Skeletons/SkeletonImage";
+import SkeletonImage from "../Skeletons/SkeletonImage";
 import { faArrowCircleLeft, faArrowCircleRight } from '@fortawesome/free-solid-svg-icons';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
@@ -94,7 +94,7 @@ const GalleryPage = () => {
                 <div className="container px-4 py-2 mx-auto md:px-6 lg:px-12 xl:px-32 mb-10">
                     {userRole === "admin" && (
                         <div className="flex flex-wrap justify-center md:justify-end mb-5">
-                            <button className="h-full bg-blue-600 text-white font-bold py-2 px-4 rounded mr-2 mb-2 md:mb-0 md:mr-0 md:ml-2 shadow-xl hover:scale-125" onClick={addImage}>
+                            <button className="h-full bg-blue-700 text-white font-bold py-2 px-4 rounded mr-2 mb-2 md:mb-0 md:mr-0 md:ml-2 shadow-xl hover:scale-125" onClick={addImage}>
                                 Add new image
                             </button>
                             
@@ -152,14 +152,14 @@ const GalleryPage = () => {
                     <div className="max-w-screen-lg mx-auto mt-5">
                         <div className="flex justify-center">
                             <button
-                                className="w-full md:w-auto h-full bg-blue-600 hover:bg-gray-700 text-white font-bold py-2 px-4 rounded ml-2 shadow-xl hover:scale-125"
+                                className="w-full md:w-auto h-full bg-blue-700 hover:bg-gray-700 text-white font-bold py-2 px-4 rounded ml-2 shadow-xl hover:scale-125"
                                 onClick={() => setCurrentPage(currentPage - 1)}
                                 disabled={startIndex === 0}
                             >
                                 <FontAwesomeIcon icon={faArrowCircleLeft} /> Previous
                             </button>
                             <button
-                                className="w-full md:w-auto h-full bg-blue-600 hover:bg-gray-700 text-white font-bold py-2 px-4 rounded ml-2 shadow-xl hover:scale-125"
+                                className="w-full md:w-auto h-full bg-blue-700 hover:bg-gray-700 text-white font-bold py-2 px-4 rounded ml-2 shadow-xl hover:scale-125"
                                 onClick={() => setCurrentPage(currentPage + 1)}
                                 disabled={endIndex >= imageUrls.length}
                             >
